+++ src/components/modals/ListModal.tsx
@@ -14,7 +14,7 @@ interface Props<T> {
   // Header Props
   title?: string
   message?: string
-  textInput?: boolean // Defaults to 'true'
+  hideSearch?: boolean // Defaults to 'false'
   initialValue?: string // Defaults to ''
   // OutlinedTextInput properties:
   searchIcon?: boolean // Defaults to 'true'
@@ -44,7 +44,7 @@ export function ListModal<T>({
   bridge,
   title,
   message,
-  textInput = true,
+  hideSearch = false,
   initialValue = '',
   rowsData = [],
   fullScreen = true,
@@ -75,7 +75,7 @@ export function ListModal<T>({
     <ThemedModal bridge={bridge} closeButton={closeArrow} onCancel={handleCancel}>
       {title == null ? null : <ModalTitle>{title}</ModalTitle>}
       {message == null ? null : <ModalMessage>{message}</ModalMessage>}
-      {textInput == null ? null : (
+      {hideSearch ? null : (
         <OutlinedTextInput
           // Our props:
           searchIcon